+++ test/empty.config.ts
@@ -1,0 +1,2 @@
+import { Configuration } from "../src/configuration";
+export const configuration: Configuration = {};