+++ CHANGELOG.md
@@ -19,9 +19,11 @@ Usability release
 
 -   Improved configuration error messages [#253][253]
 -   Provide user configuration if reporting it is valid [#251][251]
+-   Error and exit if any provided configuration is invalid [#254][254]
 
 [253]: https://github.com/atomist/automation-client-ts/issues/253
 [251]: https://github.com/atomist/automation-client-ts/issues/251
+[254]: https://github.com/atomist/automation-client-ts/issues/254
 
 ## [0.10.0][] - 2018-03-26